requirements.txt
anthropic>=0.40.0
botbuilder-core>=4.15.0
botbuilder-integration-aiohttp>=4.15.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
pydantic>=2.0.0
